backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
python-dateutil==2.9.0
pydantic==2.7.1
